Remplacement du define par un printRadix et création des utils pour les registerBit
