+++ inc/drivers/lcd_utils.h
@@ -1,0 +1,3 @@
+#include <cstdint>
+
+enum class PrintRadix : uint8_t { HEX = 16, DEC = 10, OCT = 8, BIN = 2 };
