+++ inc/drivers/rgb_lcd.h
@@ -3,16 +3,11 @@
 
 #include <inttypes.h>
 #include "STM32I2C.h"
+#include "lcd_utils.h"
 
 // Device I2C Adress
 #define LCD_ADDRESS     (0x7c)
 #define RGB_ADDRESS     (0xc4)
-
-// base define
-#define DEC 10
-#define HEX 16
-#define OCT 8
-#define BIN 2
 
 // color define 
 #define WHITE           0
@@ -81,6 +76,17 @@ private:
   void printLLNumber(uint64_t, uint8_t );
   size_t printFloat(double, uint8_t);
 
+  void send(uint8_t, uint8_t);
+  void writeRGBRegister(unsigned char addr, unsigned char dta);
+
+  uint8_t _displayfunction;
+  uint8_t _displaycontrol;
+  uint8_t _displaymode;
+
+  uint8_t _initialized;
+
+  uint8_t _numlines;
+  uint8_t _currline;
 protected:
   void setWriteError(int err = 1) { write_error = err; }
   
@@ -159,25 +165,12 @@ public:
 
   size_t print(const char[]);
   size_t print(char);
-  size_t print(unsigned char, int = DEC);
-  size_t print(int, int = DEC);
-  size_t print(unsigned int, int = DEC);
-  size_t print(long, int = DEC);
-  size_t print(unsigned long, int = DEC);
+  size_t print(unsigned char, int = static_cast<int>(PrintRadix::DEC));
+  size_t print(int, int = static_cast<int>(PrintRadix::DEC));
+  size_t print(unsigned int, int = static_cast<int>(PrintRadix::DEC));
+  size_t print(long, int = static_cast<int>(PrintRadix::DEC));
+  size_t print(unsigned long, int = static_cast<int>(PrintRadix::DEC));
   size_t print(double, int = 2);
-
-private:
-  void send(uint8_t, uint8_t);
-  void writeRGBRegister(unsigned char addr, unsigned char dta);
-
-  uint8_t _displayfunction;
-  uint8_t _displaycontrol;
-  uint8_t _displaymode;
-
-  uint8_t _initialized;
-
-  uint8_t _numlines;
-  uint8_t _currline;
 };
 }
 #endif
