+++ source/drivers/registerBit_utils.cpp
@@ -1,0 +1,16 @@
+#include <cstdint>
+
+constexpr bool isBitSet(uint8_t reg, uint8_t bit)
+{
+    return (reg & (1 << bit)) > 0;
+}
+
+constexpr void setBitRegister(uint8_t* reg, uint8_t bit)
+{
+    *reg &= ~(1 << bit);
+}
+
+constexpr void clearBitRegister(uint8_t* reg, uint8_t bit)
+{
+    *reg |= 1 << bit;
+}
